+++ Doanlesg/src/main/java/com/example/Doanlesg/services/VietQRService.java
@@ -1,6 +1,8 @@
 package com.example.Doanlesg.services;
 
 import org.springframework.stereotype.Service;
+
+import java.math.BigDecimal;
 import java.net.URLEncoder;
 import java.nio.charset.StandardCharsets;
 
@@ -10,7 +12,7 @@ public class VietQRService {
     // Thay thế bằng thông tin tài khoản của bạn
     private static final String VIETQR_TEMPLATE = "https://api.vietqr.io/image/970415-100877543376-WEMdHcP.jpg?accountName=DOLESAIGON&amount=%d&addInfo=%s";
 
-    public String generateQrCodeUrl(long amount, String description) {
+    public String generateQrCodeUrl(BigDecimal amount, String description) {
         try {
             String encodedDesc = URLEncoder.encode(description, StandardCharsets.UTF_8.name());
             return String.format(VIETQR_TEMPLATE, amount, encodedDesc);
